requirements.txt
python-dotenv
Flask
gunicorn
gspread
oauth2client
pylint
plotly
matplotlib

